requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.4
django-filter==23.2
django-redis==5.3.0
djangorestframework==3.14.0
Markdown==3.4.4
Pillow==10.0.0
python-decouple==3.8
pytz==2023.3
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
